Refactor code
Refactor code to separate Inputs from InputContainer.
